Use a do while loop to determine if user wants to play again
A do-while loop was added to keep playing as long as the user does
want to play again. A score board was also added along with a round
counter that updates after each round. A final winner message was added
at the end to display the grand winner of the game.